requirements.txt
discord.py
soundfile
numpy
discord-ext-voice-recv
dotenv
openai-whisper
torch
